kube-app/python-flask-producer/requirements.txt
Flask==2.0.1
Werkzeug==2.2.2
pandas
openpyxl
kafka-python
psycopg2-binary
